move telemetry inside sdk
