+++ src/promptflow/promptflow/_cli/_pf/entry.py
@@ -6,8 +6,7 @@ import json
 import time
 
 from promptflow._cli._utils import _get_cli_activity_name
-from promptflow._telemetry.activity import ActivityType, log_activity
-from promptflow._telemetry.telemetry import get_telemetry_logger
+from promptflow._sdk._telemetry import ActivityType, get_telemetry_logger, log_activity
 
 # Log the start time
 start_time = time.perf_counter()
