+++ src/promptflow/promptflow/_cli/_pf_azure/entry.py
@@ -7,8 +7,7 @@ import time
 
 from promptflow._cli._user_agent import USER_AGENT
 from promptflow._cli._utils import _get_cli_activity_name, get_client_info_for_cli
-from promptflow._telemetry.activity import ActivityType, log_activity
-from promptflow._telemetry.telemetry import get_telemetry_logger
+from promptflow._sdk._telemetry import ActivityType, get_telemetry_logger, log_activity
 
 # Log the start time
 start_time = time.perf_counter()
