+++ src/promptflow/promptflow/_sdk/operations/_tool_operations.py
@@ -12,7 +12,7 @@ from typing import Union
 from promptflow._core.tool_meta_generator import _parse_tool_from_function, asdict_without_none, is_tool
 from promptflow._core.tools_manager import collect_package_tools
 from promptflow._sdk._constants import ICON, ICON_DARK, ICON_LIGHT
-from promptflow._telemetry.activity import ActivityType, monitor_operation
+from promptflow._sdk._telemetry import ActivityType, monitor_operation
 from promptflow._utils.multimedia_utils import convert_multimedia_data_to_base64
 from promptflow.contracts.multimedia import Image
 from promptflow.exceptions import UserErrorException
